+++ mocks/handlers/search/data.ts
@@ -62,40 +62,9 @@ const sampleImages = [
   { url: "https://picsum.photos/236/354", width: 236, height: 354 },
 ];
 
-const sampleImages2 = [
-  { url: "https://picsum.photos/400", width: 400, height: 400 },
-  { url: "https://picsum.photos/200/300", width: 200, height: 300 },
-];
-
 export const memes = Array.from(Array(1024).keys()).map((id) => {
   const randomIndex = Math.floor(Math.random() * sampleImages.length);
   const { url, width, height } = sampleImages[randomIndex];
-  return {
-    memeId: id,
-    name: "무난한도전",
-    description: "무난한도전 무난한도전 무난한 무난한도전 무난한도전",
-    viewCount: 132,
-    shareCount: 1000,
-    createdDate: new Date().toString(),
-    modifiedDate: new Date().toString(),
-    image: {
-      images: [
-        {
-          imageId: 1,
-          imageUrl: url,
-          imageWidth: width,
-          imageHeight: height,
-        },
-      ],
-      count: 1,
-    },
-    tags: ["무한도전", "박명수"],
-  };
-});
-
-export const memeSample = Array.from(Array(1024).keys()).map((id) => {
-  const randomIndex = Math.floor(Math.random() * sampleImages2.length);
-  const { url, width, height } = sampleImages2[randomIndex];
   return {
     memeId: id,
     name: "무난한도전",
