+++ src/components/home/MemeList/MemeList.tsx
@@ -1,17 +1,35 @@
+import { useContext } from "react";
+import { css } from "twin.macro";
+
+import { useGetMemesBySort, useIntersect } from "@/application/hooks";
 import { Masonry } from "@/components/common/Masonry";
 import { MemeItem } from "@/components/meme/MemeItem";
-import type { Meme } from "@/types";
 
-interface Props {
-  memeList: Meme[];
-}
+import { MemeListContext } from "../MemeContext";
 
-export const MemeList = ({ memeList }: Props) => {
+export const MemeList = () => {
+  const sort = useContext(MemeListContext);
+  const { data: memeList, fetchNextPage } = useGetMemesBySort(sort);
+
+  const ref = useIntersect(async () => {
+    fetchNextPage();
+  });
+
   return (
-    <Masonry columns={2} spacing={9}>
-      {memeList.map((meme) => (
-        <MemeItem key={meme.memeId} meme={meme} />
-      ))}
-    </Masonry>
+    <div
+      css={[
+        css`
+          width: 100%;
+          min-height: 300px;
+        `,
+      ]}
+    >
+      <Masonry columns={2} spacing={9}>
+        {memeList.map((meme) => (
+          <MemeItem key={meme.memeId} meme={meme} />
+        ))}
+      </Masonry>
+      <div className="h-20" ref={ref} />
+    </div>
   );
 };
