+++ src/components/home/MemeList/index.ts
@@ -1,3 +1,2 @@
-export * from "./RecentMemeList";
-export * from "./SharedMemeList";
+export * from "./MemeList";
 export * from "./UserFindMemeList";
